+++ internal/handlers/tigris/fetch.go
@@ -13,63 +13,3 @@
 // limitations under the License.
 
 package tigris
-
-import (
-	"context"
-
-	"github.com/tigrisdata/tigris-client-go/driver"
-	"go.uber.org/zap"
-
-	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tigrisdb"
-	"github.com/FerretDB/FerretDB/internal/tjson"
-	"github.com/FerretDB/FerretDB/internal/types"
-	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
-)
-
-// fetch fetches all documents from the given database and collection.
-//
-// TODO https://github.com/FerretDB/FerretDB/issues/372
-func (h *Handler) fetch(ctx context.Context, param tigrisdb.FetchParam) ([]*types.Document, error) {
-	db := h.db.Driver.UseDatabase(param.DB)
-
-	collection, err := db.DescribeCollection(ctx, param.Collection)
-	switch err := err.(type) {
-	case nil:
-		// do nothing
-	case *driver.Error:
-		if tigrisdb.IsNotFound(err) {
-			h.L.Debug(
-				"Collection doesn't exist, handling a case to deal with a non-existing collection (return empty list)",
-				zap.String("db", param.DB), zap.String("collection", param.Collection),
-			)
-			return []*types.Document{}, nil
-		}
-		return nil, lazyerrors.Error(err)
-	default:
-		return nil, lazyerrors.Error(err)
-	}
-
-	var schema tjson.Schema
-	if err = schema.Unmarshal(collection.Schema); err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	iter, err := db.Read(ctx, param.Collection, driver.Filter(`{}`), nil)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-	defer iter.Close()
-
-	var res []*types.Document
-	var d driver.Document
-	for iter.Next(&d) {
-		doc, err := tjson.Unmarshal(d, &schema)
-		if err != nil {
-			return nil, lazyerrors.Error(err)
-		}
-
-		res = append(res, doc.(*types.Document))
-	}
-
-	return res, iter.Err()
-}
