+++ internal/handlers/tigris/msg_delete.go
@@ -90,7 +90,7 @@ func (h *Handler) MsgDelete(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, 
 			)
 		}
 
-		fetchedDocs, err := h.fetch(ctx, fp)
+		fetchedDocs, err := h.db.QueryDocuments(ctx, fp)
 		if err != nil {
 			return nil, err
 		}
