+++ internal/handlers/tigris/tigrisdb/tigrisdb.go
@@ -20,6 +20,7 @@ import (
 
 	"github.com/tigrisdata/tigris-client-go/config"
 	"github.com/tigrisdata/tigris-client-go/driver"
+	"go.uber.org/zap"
 
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
 )
@@ -27,16 +28,11 @@ import (
 // TigrisDB represents a Tigris database connection.
 type TigrisDB struct {
 	Driver driver.Driver
-}
-
-// FetchParam represents options/parameters used by the fetch.
-type FetchParam struct {
-	DB         string
-	Collection string
+	L      *zap.Logger
 }
 
 // New returns a new TigrisDB.
-func New(cfg *config.Driver) (*TigrisDB, error) {
+func New(cfg *config.Driver, logger *zap.Logger) (*TigrisDB, error) {
 	d, err := driver.NewDriver(context.TODO(), cfg)
 	if err != nil {
 		return nil, lazyerrors.Error(err)
@@ -44,5 +40,6 @@ func New(cfg *config.Driver) (*TigrisDB, error) {
 
 	return &TigrisDB{
 		Driver: d,
+		L:      logger,
 	}, nil
 }
